Rename and split DocumentChecker as Content/Topic
Previously the DocumentChecker covered issues about content and topics,
while the Checker class combined these issues with ImageChecker issues.

DocumentChecker sounds like it should cover everything, including
images, so this splits the existing class and renames the Checker class
to take its place.
