+++ app/views/documents/edit.html.erb
@@ -24,7 +24,7 @@
             "contextual-guidance": "document-title-guidance"
           }
         },
-        maxlength: Requirements::DocumentChecker::TITLE_MAX_LENGTH
+        maxlength: Requirements::ContentChecker::TITLE_MAX_LENGTH
       } %>
 
       <%= render "components/url_preview", {
